Move module in `utils` folder and improve some functions
